Foundational start with login and session handling.
Added command `gleshys-start` where a login procedure is executed. If
no options are provided on command line it will offer to select
organization and project for the session.

All settings are stored in a config file, by default it will be
`~/.config/gleshys.cfg` but it can be changed either using a command
line option or by setting environment variable `GLESHYSSETTINGS`.

The commands for setting current organization and project can be
executed at will, the take immidiate effect. If only one organization
(or project) is found, that is used without asking further questions -
it's made clear in STDOUT.