Support for file storage volumes.
Except working with access lists (pretty major feature as it is
absolutley needed to make the volume work for real) every bit of
adding, editing and removing a file storage volume is in place.

The access list is missing as I have not touched the server list
feature yet. It's dead simple to add the sole thing but sticking the
fingers into the server end point is a real time sink at this
moment. Use the web client or write your own accesslist magic in the
mean time. It's just a comma separated list of server keys if I recall
correct.