BUG: Use provided settings file when stepping outside of script.
When asking a fellow script for information the settings file must
always be passed on command line. It is impossible to know if the
environment is tainted or not.

In this case `file-storage edit` is asking `server list` about what
servers are available. If the user provided a different file in
environment/home directory than on command line to `bin/file-storage`
then `bin/server` would try to read invalid data.