FEATURE: Support delayed stop/reboot of servers.
The delayed reboot or stopping of a server is not supported on KVM,
the action will be immediate.
For VM Ware you need to have vmware-tools installed to use the feature.

This minor feature is not particular well supported it seems. You are
probably better of using a sleep command to do it.

  $ sleep 500 && server stop wps12345