FEATURE: Add shorthand for adding or removing servers from volume access.
The volume access list was onlu available via the edit command earlier.

 $ gleshys file-storage edit -l key1,key2,..keyN fs-volumeKey

Having short hand functions for adding and removing servers from a
volume access list is a neat idea.

 $ gleshys file-storage grant key1 fs-volumeKey
 $ gleshys file-storage grant key1,key2 fs-volumeKey
 $ gleshys file-storage revoke key2 fs-volumeKey
 $ gleshys file-storage revoke all fs-volumeKey

The edit feature still can manipulate the list but as before it always
replaces the complete list with what is provided.

The internal edit function no longer demand the auto promotion setting
to be passed. The grant and revoke functions do not know about that
value and wouldn't be able to reuse the internal edit function
otherwise.